+++ src/main/java/com/example/FifthSpringFinal/dto/ParametersDTO.java
@@ -1,13 +1,14 @@
 package com.example.FifthSpringFinal.dto;
 
 import javax.persistence.*;
+import javax.persistence.criteria.CriteriaBuilder;
 
 @Entity
 @Table(name = "parameters")
 public class ParametersDTO {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Integer id;
 
     @Column(name="num_of_shares", nullable = false)
     private int numOfShares;
